requirements.txt
torch==2.1.2
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors>=0.4.2
Pillow
numpy
pyyaml
python-dotenv
requests
moviepy
opencv-python
git+https://github.com/suno-ai/bark.git
scipy